Drivers. HTU21D for NRF51. I2C read function seems to work as expected. I need to pass the TW0 object by reference to improve it

+++ Drivers/HTU21D/HTU21D.h
@@ -99,4 +99,4 @@ uint32_t  HTU21D_TriggerTemperature ( void );
 uint32_t  HTU21D_ReadTemperature    ( void );
 uint32_t  HTU21D_TriggerHumidity    ( void );
 uint32_t  HTU21D_ReadHumidity       ( void );
-uint32_t  HTU21D_BatteryStatus      ( uint32_t* battStatus );
+uint32_t  HTU21D_BatteryStatus      ( uint8_t* battStatus );
